+++ src/main/java/com/ssw/epicgames/controlleradvice/GlobalControllerAdvice.java
@@ -1,15 +1,8 @@
 package com.ssw.epicgames.controlleradvice;
 
-
 import com.ssw.epicgames.entities.UserEntity;
-import com.ssw.epicgames.services.PageService;
 import jakarta.servlet.http.HttpSession;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.ModelAndViewDefiningException;
